refactor(ui): extract message-list reconciliation seam
